refactor(LC027): split foreign-key analysis helpers
- separate entity checks, configuration scanning, and type lookup logic
- keep analyzer behavior unchanged while reducing local complexity
